Move new method to existing LongUtils class in oak-commons.
