+++ oak-store-spi/src/main/java/org/apache/jackrabbit/oak/json/JsonDeserializer.java
@@ -27,6 +27,7 @@ import javax.jcr.PropertyType;
 import org.apache.jackrabbit.guava.common.collect.Lists;
 import org.apache.jackrabbit.oak.api.PropertyState;
 import org.apache.jackrabbit.oak.api.Type;
+import org.apache.jackrabbit.oak.commons.LongUtils;
 import org.apache.jackrabbit.oak.commons.json.JsopReader;
 import org.apache.jackrabbit.oak.commons.json.JsopTokenizer;
 import org.apache.jackrabbit.oak.plugins.memory.BinaryPropertyState;
@@ -113,7 +114,7 @@ public class JsonDeserializer {
     private PropertyState readProperty(String name, JsopReader reader) {
         if (reader.matches(JsopReader.NUMBER)) {
             String number = reader.getToken();
-            Long maybeLong = ParseUtils.tryParse(number);
+            Long maybeLong = LongUtils.tryParse(number);
             if (maybeLong == null) {
                 return new DoublePropertyState(name, Double.parseDouble(number));
             } else {
@@ -170,7 +171,7 @@ public class JsonDeserializer {
         while (!reader.matches(']')) {
             if (reader.matches(JsopReader.NUMBER)) {
                 String number = reader.getToken();
-                Long maybeLong = ParseUtils.tryParse(number);
+                Long maybeLong = LongUtils.tryParse(number);
                 if (maybeLong == null) {
                     type = PropertyType.DOUBLE;
                     values.add(Double.parseDouble(number));
